+++ src/shaders/mod.rs
@@ -4,7 +4,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-mod conical_gradient;
 mod gradient;
 mod linear_gradient;
 mod pattern;
@@ -12,7 +11,6 @@ mod radial_gradient;
 
 use tiny_skia_path::{NormalizedF32, Scalar};
 
-pub use conical_gradient::ConicalGradient;
 pub use gradient::GradientStop;
 pub use linear_gradient::LinearGradient;
 pub use pattern::{FilterQuality, Pattern, PixmapPaint};
@@ -52,8 +50,8 @@ pub enum Shader<'a> {
     SolidColor(Color),
     /// A linear gradient shader.
     LinearGradient(LinearGradient),
-    /// A conical gradient shader.
-    ConicalGradient(ConicalGradient),
+    /// A conical radial gradient shader.
+    RadialGradient(RadialGradient),
     /// A pattern shader.
     Pattern(Pattern<'a>),
 }
@@ -64,7 +62,7 @@ impl Shader<'_> {
         match self {
             Shader::SolidColor(ref c) => c.is_opaque(),
             Shader::LinearGradient(ref g) => g.is_opaque(),
-            Shader::ConicalGradient(_) => false,
+            Shader::RadialGradient(_) => false,
             Shader::Pattern(_) => false,
         }
     }
@@ -81,7 +79,7 @@ impl Shader<'_> {
                 true
             }
             Shader::LinearGradient(ref g) => g.push_stages(cs, p),
-            Shader::ConicalGradient(ref g) => g.push_stages(cs, p),
+            Shader::RadialGradient(ref g) => g.push_stages(cs, p),
             Shader::Pattern(ref patt) => patt.push_stages(cs, p),
         }
     }
@@ -93,7 +91,7 @@ impl Shader<'_> {
             Shader::LinearGradient(g) => {
                 g.base.transform = g.base.transform.post_concat(ts);
             }
-            Shader::ConicalGradient(g) => {
+            Shader::RadialGradient(g) => {
                 g.base.transform = g.base.transform.post_concat(ts);
             }
             Shader::Pattern(p) => {
@@ -123,7 +121,7 @@ impl Shader<'_> {
             Shader::LinearGradient(g) => {
                 g.base.apply_opacity(opacity);
             }
-            Shader::ConicalGradient(g) => {
+            Shader::RadialGradient(g) => {
                 g.base.apply_opacity(opacity);
             }
             Shader::Pattern(ref mut p) => {